+++ command_line_assistant/__main__.py
@@ -1,37 +1,22 @@
-import argparse
 import sys
 
-from command_line_assistant.commands import add_default_command
-from command_line_assistant.commands.history import HistoryCommand
-from command_line_assistant.commands.query import QueryCommand
-
-
-def create_argument_parser() -> argparse.ArgumentParser:
-    """Create the argument parser for command line assistant."""
-    parser = argparse.ArgumentParser(
-        description="A script with multiple optional arguments and a required positional argument if no optional arguments are provided.",
-    )
-    parser.add_argument(
-        "--version",
-        action="version",
-        version="0.1.0",
-        default=argparse.SUPPRESS,
-        help="Show command line assistant version and exit.",
-    )
-    commands_parser = parser.add_subparsers(
-        dest="command", help="command line assistant helpers"
-    )
-
-    QueryCommand.register_subcommand(commands_parser)  # type: ignore
-    HistoryCommand.register_subcommand(commands_parser)  # type: ignore
-
-    return parser
+from command_line_assistant.commands import history, query
+from command_line_assistant.commands.utils import (
+    add_default_command,
+    create_argument_parser,
+)
 
 
 def main() -> int:
-    parser = create_argument_parser()
+    parser, commands_parser = create_argument_parser()
 
-    args = add_default_command(sys.argv[1:])
+    # TODO: add autodetection of BaseCLICommand classes in the future so we can just drop
+    # new subcommand python modules into the directory and then loop and call `register_subcommand()`
+    # on each one.
+    query.register_subcommand(commands_parser)  # type: ignore
+    history.register_subcommand(commands_parser)  # type: ignore
+
+    args = add_default_command(sys.argv)
     args = parser.parse_args(args)
 
     if not hasattr(args, "func"):
